Simplify shopping list scripts and merge Google Sheets sync
