--- Assets/1-Scripts/ShoppingList/GoogleSheetsShoppingListWriter.cs.meta
@@ -1,22 +1,0 @@
-fileFormatVersion: 2
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
-guid: a5370bcbadaa600409546f64c548d1bf
-=======
-guid: 30db58e368d738344b5fbb23a616a309
->>>>>>> Stashed changes
-=======
-guid: 30db58e368d738344b5fbb23a616a309
->>>>>>> Stashed changes
-=======
-guid: 30db58e368d738344b5fbb23a616a309
->>>>>>> Stashed changes
-=======
-guid: 30db58e368d738344b5fbb23a616a309
->>>>>>> Stashed changes
-=======
-guid: 30db58e368d738344b5fbb23a616a309
->>>>>>> Stashed changes
