+++ Assets/3-Scenes/Main.unity
@@ -1672,19 +1672,7 @@ MonoBehaviour:
           m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
           m_IntArgument: 0
           m_FloatArgument: 0
-          m_StringArgument: 
-          m_BoolArgument: 0
-        m_CallState: 2
-      - m_Target: {fileID: 281186835}
-        m_TargetAssemblyTypeName: GoogleSheetsShoppingListWriter, Assembly-CSharp
-        m_MethodName: UploadList
-        m_Mode: 2
-        m_Arguments:
-          m_ObjectArgument: {fileID: 281186833}
-          m_ObjectArgumentAssemblyTypeName: ShoppingListManager, Assembly-CSharp
-          m_IntArgument: 0
-          m_FloatArgument: 0
-          m_StringArgument: 
+          m_StringArgument:
           m_BoolArgument: 0
         m_CallState: 2
 --- !u!114 &227690584
@@ -1732,7 +1720,6 @@ GameObject:
   m_Component:
   - component: {fileID: 281186831}
   - component: {fileID: 281186836}
-  - component: {fileID: 281186835}
   - component: {fileID: 281186834}
   - component: {fileID: 281186833}
   m_Layer: 0
@@ -1786,21 +1773,7 @@ MonoBehaviour:
   itemContainer: {fileID: 3973478338881549880}
   completedItemContainer: {fileID: 1990300659}
   itemPrefab: {fileID: 6455409839748775875, guid: 70fdb8f68f3242043ab9424243209970, type: 3}
---- !u!114 &281186835
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 281186829}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a5370bcbadaa600409546f64c548d1bf, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  scriptUrl: https://script.google.com/macros/s/AKfycbwt2gxR-pnYT3X4dpeMmpdG4vH2GS7ymy8lVwWOEZrYvBEivFLkiNOvYsnzpoYyCoTv/exec
-  manager: {fileID: 281186833}
---- !u!114 &281186836
+  --- !u!114 &281186836
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -1814,13 +1787,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   manager: {fileID: 281186833}
   sheetUrl: https://docs.google.com/spreadsheets/d/e/2PACX-1vQOPzajguiKaAMJizdIgBxxXgQ-N1dmEH8-tnbo_Rx_LG8yByPp2CVKd_12UlrZyOtzNdVdm4b8FgNG/pub?output=csv
-  listHeader: List
-  itemHeader: Item
-  quantityHeader: Units
-  positionHeader: Position
-  completedHeader: Completed
-  idHeader: Id
-  defaultListName: List
+  scriptUrl: https://script.google.com/macros/s/AKfycbwt2gxR-pnYT3X4dpeMmpdG4vH2GS7ymy8lVwWOEZrYvBEivFLkiNOvYsnzpoYyCoTv/exec
   refreshInterval: 0
 --- !u!1 &282628623
 GameObject:
@@ -9060,19 +9027,7 @@ MonoBehaviour:
           m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
           m_IntArgument: 0
           m_FloatArgument: 0
-          m_StringArgument: 
-          m_BoolArgument: 0
-        m_CallState: 2
-      - m_Target: {fileID: 281186835}
-        m_TargetAssemblyTypeName: GoogleSheetsShoppingListWriter, Assembly-CSharp
-        m_MethodName: UploadList
-        m_Mode: 2
-        m_Arguments:
-          m_ObjectArgument: {fileID: 281186833}
-          m_ObjectArgumentAssemblyTypeName: ShoppingListManager, Assembly-CSharp
-          m_IntArgument: 0
-          m_FloatArgument: 0
-          m_StringArgument: 
+          m_StringArgument:
           m_BoolArgument: 0
         m_CallState: 2
 --- !u!114 &1562326741
